pyproject.toml
[tool.poetry]
name = "prefect2-flows"
version = "0.1.0"
description = ""
authors = ["Matthias Schlögl <matthias.schloegl@oeaw.ac.at>"]
license = "MIT"
packages = [
    { include = "prefect2-flows", from="." },
]

[tool.poetry.dependencies]
python = "^3.11"
httpx = "^0.25.0"
gitpython = "^3.1.37"
requests = "^2.31.0"
rdflib = "^7.0.0"
sparqlwrapper = "^2.0.0"
deepl = "^1.15.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
